chore: rename pages to views
